+++ eslint.config.js
@@ -2,7 +2,7 @@ import js from '@eslint/js'
 import tseslint from 'typescript-eslint'
 
 export default tseslint.config(
-  { ignores: ['out/**', 'dist/**', 'node_modules/**'] },
+  { ignores: ['out/**', 'dist/**', 'node_modules/**', 'scripts/**'] },
   js.configs.recommended,
   ...tseslint.configs.recommended,
   {